commit cd9341501a
    15205 SCardBeginTransaction should probably block
commit 494aa9a4ca
    15960 ccidadm should use ilstr when building strings
commit 60afb9d1b4
    16512 xhci leaks oneshot interrupt IN transfers on pipe reset
commit bf20feb33f
    16605 lmrc smatch issues
commit 7001e2dfbd
    16607 libnvme NSID field check incorrect for get-logpage
commit 1a90c98d75
    16319 tools: add yacc to build tools
commit 63df5b3228
    16612 libshare does not build with libxml2 2.13
commit 71536d92c2
    16573 smntemp Tctl is not degrees C
    16574 smntemp CCD Tdie support
    16575 ksensor/topo support for synthetic sensors
    16576 want ksensor manuals
    16577 libtopo sensor property translations truncate values
commit f1cd2b6cd0
    16076 pwdx should work on core files
    16553 want makedev variant for full 32-bit dev_t
    16554 libproc default ops should set errno

Conflicts:
	usr/src/uts/intel/sys/amdzen/smn.h

+++ usr/src/lib/fm/topo/libtopo/common/topo_subr.c
@@ -24,7 +24,7 @@
  */
 /*
  * Copyright 2020 Joyent, Inc.
- * Copyright 2023 Oxide Computer Company
+ * Copyright 2024 Oxide Computer Company
  */
 
 #include <alloca.h>
@@ -376,7 +376,7 @@ topo_sensor_type_name(uint32_t type, char *buf, size_t len)
 }
 
 void
-topo_sensor_units_name(uint8_t type, char *buf, size_t len)
+topo_sensor_units_name(uint32_t type, char *buf, size_t len)
 {
 	topo_name_trans_t *ntp;
 
@@ -391,7 +391,7 @@ topo_sensor_units_name(uint8_t type, char *buf, size_t len)
 }
 
 void
-topo_led_type_name(uint8_t type, char *buf, size_t len)
+topo_led_type_name(uint32_t type, char *buf, size_t len)
 {
 	topo_name_trans_t *ntp;
 
@@ -406,7 +406,7 @@ topo_led_type_name(uint8_t type, char *buf, size_t len)
 }
 
 void
-topo_led_state_name(uint8_t type, char *buf, size_t len)
+topo_led_state_name(uint32_t type, char *buf, size_t len)
 {
 	topo_name_trans_t *ntp;
 
@@ -421,7 +421,7 @@ topo_led_state_name(uint8_t type, char *buf, size_t len)
 }
 
 void
-topo_sensor_state_name(uint32_t sensor_type, uint8_t state, char *buf,
+topo_sensor_state_name(uint32_t sensor_type, uint32_t state, char *buf,
     size_t len)
 {
 	topo_name_trans_t *ntp;
